Fix and readme updatet

+++ app/Http/Controllers/MedicationController.php
@@ -18,12 +18,22 @@ class MedicationController extends Controller
 
     public function update(Request $request, $id) {
         $medication = Medication::find($id);
+
+        if(!$medication) {
+            return response()->json("Not Found", 404);
+        }
+
         $medication->update($request->all());
         return response()->json($medication, 200);
     }
 
     public function delete($id) {
         $medication = Medication::find($id);
+
+        if(!$medication) {
+            return response()->json("Not Found", 404);
+        }
+        
         $medication->delete();
         return response()->json(null, 204);
     }
